Moved getPath related methods to new PathFinder class
